:sparkles: feat: Add entity and relationship extraction functionality
This commit introduces the functionality of extracting entities and relationships from a text chunk. This is achieved by implementing the 'extract' method in the 'EntityRelationExtractor' class. The method utilizes a language model to perform entity and relationship extraction and assigns unique IDs to entities. Additionally, the IDs of relationships are updated accordingly. The 'ExtractionResult' class is also added to represent the result of entity and relationship extraction.